# docs/api-contract-project.yaml
openapi: 3.0.0
info:
  title: API Contract - Sistem Penulisan Hibah Buku (Full Project)
  description: |
    API Master Contract for all 4 modules in the Book Grant Writing System.
    (Kontrak Master API untuk ke-4 modul pada Sistem Penulisan Hibah Buku.)

    Termasuk: Modul 1 Fondasi Sistem, User, Role & Kontrak (Kelompok 4)
    Modul 2 Penulis & Manajemen Naskah (Kelompok 3)
    Modul 3 Reviewer & Proses Review (Kelompok 2)
    Modul 4 Penerbit, Pra-Cetak & Monitoring (Kelompok 1)
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local Development Server / Server Pengembangan Lokal
...
paths:
...
  /reminder/trigger:
    post:
      summary: Trigger Reminder Manually
      description: |
        Menjalankan cron reminder secara manual.
        Membaca tabel `deadlines` untuk mengirim pengingat.
        (Catatan: Skema SQL tidak menyediakan tabel reminder_logs; reminder tetap dijalankan berdasarkan deadlines.)
      tags:
        - Modul 4 - Penerbit, Pra-Cetak, & Monitoring (Kelompok 1)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reminder berhasil diproses.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reminders_sent:
                        type: integer
                      reminders_failed:
                        type: integer
                  _links:
                    type: object
                    properties:
                      dashboard:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/api/admin/dashboard-stats"
                          method:
                            type: string
                            example: GET
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Unauthenticated."
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Access denied."
...